Add configuration for the git command line application
`git` is a free and open source distributed version control system
designed to handle everything from small to very large projects with
speed and efficiency.

The configurations included in this commit extend the default
functionality of `git`, providing aliases for common use cases;
alterations to core features; a template for commit messages; a global
set of files and directories to ignore; and global file attributes.