Add configuration for the Z shell
The Z shell (`zsh`) is a Unix shell that can be used as an interactive
login shell and as a command interpreter for shell scripting. Z shell is
an extended Bourne shell with many improvements, including some features
of `bash`, `ksh`, and `tcsh`.

This commit includes configurations to extend the default functionality
of Z shell, souring additional configurations originally implemented for
Bash.