Add configuration for the vim text editor
`vim` is a highly configurable text editor built to make creating and
changing any kind of text very efficient. It is included as `vi` with
most UNIX systems and with macOS.

`vim` is rock stable and is continuously being developed to become even
better. Among its features are:

 - persistent, multi-level undo tree,
 - extensive plugin system,
 - support for hundreds of programming languages and file formats,
 - powerful search and replace,
 - integrates with many tools.

This commit includes configurations to extend the base functionality of
`vim`, including a plugin system that is updated when chezmoi is applied.