Add configuration for the tmux terminal multiplexer
`tmux` is an open-source terminal multiplexer for Unix-like operating
systems. It allows multiple terminal sessions to be accessed
simultaneously in a single window. It is useful for running more than
one command-line program at the same time.

`tmux` is used commonly with `vim` and as such much of their
configurations have been carefully crafted to work together.

This commit includes configurations to extend the default functionality
of `tmux`, including a plugin system that is updated when chezmoi is
applied.