Add configuration for the interactive Python shell
This file is executed when the Python interactive shell is started if
`$PYTHONSTARTUP` is in your environment and points to this file. It
contains regular Python statements and as such may contain any code that
would be written in any Python application. The configuration for GNU
readline, typically `~/.inputrc`, may greatly complement this file.