Add root editor configuration
EditorConfig helps maintain consistent coding styles for multiple
developers working on the same project across various editors and IDEs.
The EditorConfig project consists of a file format for defining coding
styles and a collection of text editor plugins that enable editors to
read the file format and adhere to defined styles. EditorConfig files
are easily readable and they work nicely with version control systems.

This particular configuration enforces a consistent style across all
file types, with exception to Markdown and diffs. As this file is marked
as being `root`, if no project specific configuration is given then the
rules defined in this file will be used.