Add configuration for the wget command line application
GNU `wget` is a free software package for retrieving files using HTTP,
HTTPS, FTP and FTPS, the most widely used Internet protocols. It is
a non-interactive commandline tool, so it may easily be called from
scripts, cron jobs, terminals without X-Windows support, etc.

`wget` has many features to make retrieving large files or mirroring
entire web or FTP sites easy, including:

 - can resume aborted downloads, using REST and RANGE,
 - can use filename wild cards and recursively mirror directories,
 - NLS-based message files for many different languages,
 - optionally converts absolute links in downloaded documents to
   relative, so that downloaded documents may link to each other
   locally,
 - runs on most UNIX-like operating systems as well as Microsoft
   Windows,
 - supports HTTP proxies,
 - supports HTTP cookies,
 - supports persistent HTTP connections,
 - unattended / background operation,
 - uses local file timestamps to determine whether documents need to be
   re-downloaded when mirroring.

In the majority of cases it is preferred to use `cURL` as it has support
for many additional protocols, but there are a few instances when `wget`
is more appropriate, and some scripts will favour the use of `wget`
over of `cURL`.