Add configuration for the ack command line application
`ack` is a grep-like source code search tool. Designed for programmers
with large heterogeneous trees of source code, `ack` is written in
portable Perl 5 and takes advantage of the power of Perl's regular
expressions.

This commit defines a list of file extensions and directories `ack`
should ignore when searching, and an additional set of file types `ack`
should recognise.